refactor(direct-control): extract progression closeout sources
Move the embedded technology and culture chooser closeout sources into packages/civ7-direct-control/src/play/progression/technology.ts and culture.ts. Keep the public request wrappers and build-command composition in index.ts for the later composition pass.

Record the partial operation/progression source-owner status in the support modularization OpenSpec workstream. This is local package source relocation proof only, not runtime or live-game proof.

Validation: bun run --cwd packages/civ7-direct-control test -- technology-choice-closeout.test.ts culture-choice-closeout.test.ts; bun run --cwd packages/civ7-direct-control test; bun run --cwd packages/civ7-direct-control check; bun run --cwd packages/civ7-direct-control build; bun run openspec -- validate civ7-support-direct-control-modularization --strict; bun run resources:status; git diff --check.
