+++ packages/civ7-direct-control/src/index.ts
@@ -24,7 +24,9 @@ import {
 import { notificationDismissalSource } from "./play/notifications/dismissal.js";
 import { playNotificationViewSource } from "./play/notifications/view.js";
 import { operationRouterSource } from "./play/operations/router.js";
+import { cultureChoiceCloseoutSource } from "./play/progression/culture.js";
 import { progressDashboardSource } from "./play/progression/progress-dashboard.js";
+import { technologyChoiceCloseoutSource } from "./play/progression/technology.js";
 import { readyCityViewSource } from "./play/ready/city.js";
 import { unitMovePreviewSource } from "./play/ready/move-preview.js";
 import { readyUnitViewSource } from "./play/ready/unit.js";
@@ -5772,208 +5774,6 @@ function autoplaySetterSource(options: Civ7AutoplayOptions): string {
   if (options.returnAsPlayer !== undefined) statements.push(`Autoplay.setReturnAsPlayer(${jsLiteral(options.returnAsPlayer)});`);
   if (options.pause !== undefined) statements.push(`Autoplay.setPause(${jsLiteral(options.pause)});`);
   return statements.join("\n    ");
-}
-
-function technologyChoiceCloseoutSource(): string {
-  return `${probeHelperSource()}
-    const toComponentId = (value) => {
-      if (!value || typeof value !== "object") return null;
-      if (typeof value.owner !== "number" || typeof value.id !== "number") return null;
-      const out = { owner: value.owner, id: value.id };
-      if (typeof value.type === "number") out.type = value.type;
-      return out;
-    };
-    const safeCall = (label, fn) => {
-      try {
-        return { ok: true, value: fn() };
-      } catch (err) {
-        return { ok: false, error: label + ": " + String(err) };
-      }
-    };
-    const successFromCanStart = (value) => value?.Success === true || value?.canStart === true;
-    const readTechnologyState = (playerId) => ({
-      currentResearching: probe(() => Players.get(playerId)?.Techs?.getResearching?.() ?? null),
-      targetNode: probe(() => Players.get(playerId)?.Techs?.getTargetNode?.() ?? null),
-    });
-    const currentTechnologyNotification = () => safeCall("find current technology-choice notification", () => {
-      const ids = typeof Game.Notifications.getIdsForPlayer === "function"
-        ? Game.Notifications.getIdsForPlayer(GameContext.localPlayerID)
-        : [];
-      const rows = Array.isArray(ids) ? ids : [];
-      for (const id of rows) {
-        const type = typeof Game.Notifications.getType === "function"
-          ? Game.Notifications.getType(id)
-          : Game.Notifications.find(id)?.Type;
-        const typeName = typeof Game.Notifications.getTypeName === "function" ? Game.Notifications.getTypeName(type) : null;
-        if (String(typeName ?? "").toUpperCase().includes("CHOOSE_TECH")) return toComponentId(id);
-      }
-      return null;
-    });
-    const sendTechnologyChoiceCloseout = (input) => {
-      const localPlayerId = GameContext.localPlayerID;
-      const playerId = Number.isInteger(input.playerId) ? input.playerId : localPlayerId;
-      const node = Number(input.node);
-      const selectedNotification = toComponentId(input.notificationId) ?? currentTechnologyNotification().value ?? null;
-      const beforeTechnology = readTechnologyState(playerId);
-      const activationResult = input.activateNotification === false
-        ? { ok: false, skipped: true, reason: "activation disabled" }
-        : safeCall("Game.Notifications.activate", () => selectedNotification ? Game.Notifications.activate(selectedNotification) : null);
-      const chooseArgs = { ProgressionTreeNodeType: node };
-      const noNode = typeof ProgressionTreeNodeTypes !== "undefined" && typeof ProgressionTreeNodeTypes.NO_NODE === "number"
-        ? ProgressionTreeNodeTypes.NO_NODE
-        : -1;
-      const clearArgs = { ProgressionTreeNodeType: noNode };
-      const canChoose = safeCall("Game.PlayerOperations.canStart SET_TECH_TREE_NODE", () => Game.PlayerOperations.canStart(
-        playerId,
-        PlayerOperationTypes.SET_TECH_TREE_NODE,
-        chooseArgs,
-        false,
-      ));
-      const chooseResult = canChoose.ok && successFromCanStart(canChoose.value)
-        ? safeCall("Game.PlayerOperations.sendRequest SET_TECH_TREE_NODE", () => Game.PlayerOperations.sendRequest(
-          playerId,
-          PlayerOperationTypes.SET_TECH_TREE_NODE,
-          chooseArgs,
-        ))
-        : { ok: false, skipped: true, reason: "SET_TECH_TREE_NODE did not validate" };
-      const canClearTarget = safeCall("Game.PlayerOperations.canStart SET_TECH_TREE_TARGET_NODE", () => Game.PlayerOperations.canStart(
-        playerId,
-        PlayerOperationTypes.SET_TECH_TREE_TARGET_NODE,
-        clearArgs,
-        false,
-      ));
-      const clearTargetResult = canClearTarget.ok && successFromCanStart(canClearTarget.value)
-        ? safeCall("Game.PlayerOperations.sendRequest SET_TECH_TREE_TARGET_NODE", () => Game.PlayerOperations.sendRequest(
-          playerId,
-          PlayerOperationTypes.SET_TECH_TREE_TARGET_NODE,
-          clearArgs,
-        ))
-        : { ok: false, skipped: true, reason: "SET_TECH_TREE_TARGET_NODE did not validate" };
-      return {
-        localPlayerId,
-        playerId,
-        node,
-        notificationId: selectedNotification,
-        beforeTechnology,
-        activationResult,
-        canChoose,
-        chooseResult,
-        canClearTarget,
-        clearTargetResult,
-        afterTechnology: readTechnologyState(playerId),
-        sent: chooseResult.ok === true && clearTargetResult.ok === true,
-        notes: [
-          "This uses the App UI owner for technology chooser closeout: optional Game.Notifications.activate, SET_TECH_TREE_NODE, then SET_TECH_TREE_TARGET_NODE with NO_NODE.",
-          "The caller must still re-read notification state; successful App UI sends are not proof that the technology-choice blocker cleared."
-        ],
-      };
-    };`;
-}
-
-function cultureChoiceCloseoutSource(): string {
-  return `${probeHelperSource()}
-    const toComponentId = (value) => {
-      if (!value || typeof value !== "object") return null;
-      if (typeof value.owner !== "number" || typeof value.id !== "number") return null;
-      const out = { owner: value.owner, id: value.id };
-      if (typeof value.type === "number") out.type = value.type;
-      return out;
-    };
-    const safeCall = (label, fn) => {
-      try {
-        return { ok: true, value: fn() };
-      } catch (err) {
-        return { ok: false, error: label + ": " + String(err) };
-      }
-    };
-    const successFromCanStart = (value) => value?.Success === true || value?.canStart === true;
-    const readCultureState = (playerId) => ({
-      currentResearching: probe(() => {
-        const culture = Players.get(playerId)?.Culture;
-        const activeTree = culture?.getActiveTree?.();
-        if (activeTree == null) return null;
-        const tree = Game.ProgressionTrees.getTree(playerId, activeTree);
-        const activeNodeIndex = tree?.activeNodeIndex;
-        return Number.isFinite(Number(activeNodeIndex)) && activeNodeIndex >= 0
-          ? tree?.nodes?.[activeNodeIndex]?.nodeType ?? null
-          : null;
-      }),
-      targetNode: probe(() => Players.get(playerId)?.Culture?.getTargetNode?.() ?? null),
-      availableNodeTypes: probe(() => Players.get(playerId)?.Culture?.getAllAvailableNodeTypes?.() ?? []),
-    });
-    const currentCultureNotification = () => safeCall("find current culture-choice notification", () => {
-      const ids = typeof Game.Notifications.getIdsForPlayer === "function"
-        ? Game.Notifications.getIdsForPlayer(GameContext.localPlayerID)
-        : [];
-      const rows = Array.isArray(ids) ? ids : [];
-      for (const id of rows) {
-        const type = typeof Game.Notifications.getType === "function"
-          ? Game.Notifications.getType(id)
-          : Game.Notifications.find(id)?.Type;
-        const typeName = typeof Game.Notifications.getTypeName === "function" ? Game.Notifications.getTypeName(type) : null;
-        if (String(typeName ?? "").toUpperCase().includes("CHOOSE_CULTURE")) return toComponentId(id);
-      }
-      return null;
-    });
-    const sendCultureChoiceCloseout = (input) => {
-      const localPlayerId = GameContext.localPlayerID;
-      const playerId = Number.isInteger(input.playerId) ? input.playerId : localPlayerId;
-      const node = Number(input.node);
-      const selectedNotification = toComponentId(input.notificationId) ?? currentCultureNotification().value ?? null;
-      const beforeCulture = readCultureState(playerId);
-      const activationResult = input.activateNotification === false
-        ? { ok: false, skipped: true, reason: "activation disabled" }
-        : safeCall("Game.Notifications.activate", () => selectedNotification ? Game.Notifications.activate(selectedNotification) : null);
-      const chooseArgs = { ProgressionTreeNodeType: node };
-      const noNode = typeof ProgressionTreeNodeTypes !== "undefined" && typeof ProgressionTreeNodeTypes.NO_NODE === "number"
-        ? ProgressionTreeNodeTypes.NO_NODE
-        : -1;
-      const clearArgs = { ProgressionTreeNodeType: noNode };
-      const canChoose = safeCall("Game.PlayerOperations.canStart SET_CULTURE_TREE_NODE", () => Game.PlayerOperations.canStart(
-        playerId,
-        PlayerOperationTypes.SET_CULTURE_TREE_NODE,
-        chooseArgs,
-        false,
-      ));
-      const chooseResult = canChoose.ok && successFromCanStart(canChoose.value)
-        ? safeCall("Game.PlayerOperations.sendRequest SET_CULTURE_TREE_NODE", () => Game.PlayerOperations.sendRequest(
-          playerId,
-          PlayerOperationTypes.SET_CULTURE_TREE_NODE,
-          chooseArgs,
-        ))
-        : { ok: false, skipped: true, reason: "SET_CULTURE_TREE_NODE did not validate" };
-      const canClearTarget = safeCall("Game.PlayerOperations.canStart SET_CULTURE_TREE_TARGET_NODE", () => Game.PlayerOperations.canStart(
-        playerId,
-        PlayerOperationTypes.SET_CULTURE_TREE_TARGET_NODE,
-        clearArgs,
-        false,
-      ));
-      const clearTargetResult = canClearTarget.ok && successFromCanStart(canClearTarget.value)
-        ? safeCall("Game.PlayerOperations.sendRequest SET_CULTURE_TREE_TARGET_NODE", () => Game.PlayerOperations.sendRequest(
-          playerId,
-          PlayerOperationTypes.SET_CULTURE_TREE_TARGET_NODE,
-          clearArgs,
-        ))
-        : { ok: false, skipped: true, reason: "SET_CULTURE_TREE_TARGET_NODE did not validate" };
-      return {
-        localPlayerId,
-        playerId,
-        node,
-        notificationId: selectedNotification,
-        beforeCulture,
-        activationResult,
-        canChoose,
-        chooseResult,
-        canClearTarget,
-        clearTargetResult,
-        afterCulture: readCultureState(playerId),
-        sent: chooseResult.ok === true && clearTargetResult.ok === true,
-        notes: [
-          "This uses the App UI owner for culture chooser closeout: optional Game.Notifications.activate, SET_CULTURE_TREE_NODE, then SET_CULTURE_TREE_TARGET_NODE with NO_NODE.",
-          "The caller must still re-read notification state; successful App UI sends are not proof that the culture-choice blocker cleared."
-        ],
-      };
-    };`;
 }
 
 function diplomacyResponseCloseoutSource(): string {
